requirements.txt
Flask>=2.0
google-generativeai>=0.5.0  # Google's SDK for Gemini
python-dotenv>=0.19.0
gunicorn
requests